requirements.txt
pandas 
numpy
matplotlib
langchain
langchain-community
langchain-experimental
langchain-openai
langgraph
python-dotenv
fastapi
uvicorn
azure-identity
pyodbc
